CT-179 사용자 관리 페이지 분리
