+++ src/app/(admin)/admin/(management)/user/layout.tsx
@@ -1,0 +1,9 @@
+import AdminUserLayoutContainer from "@/app/(admin)/_components/AdminUserLayoutContainer";
+
+export default function AdminUserManagementLayout({
+    children,
+  }: Readonly<{
+    children: React.ReactNode;
+  }>) {
+    return <AdminUserLayoutContainer>{children}</AdminUserLayoutContainer>;
+  }
